requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Login
python-dotenv
Werkzeug  # Normalerweise schon bei Flask dabei, aber explizit ist sicher
email_validator # Wird von Flask-WTF für Email-Validierung benötigt